+++ RUFAS/biophysical/manure/storage/anaerobic_lagoon.py
@@ -48,6 +48,11 @@ class AnaerobicLagoon(Storage):
             capacity=capacity,
         )
 
+    @property
+    def _emptying_fraction(self) -> float:
+        """The fraction of manure to be emptied from the Anaerobic Lagoon when it is time to empty it."""
+        return 1.0 - ManureConstants.ANAEROBIC_LAGOON_MANURE_RETENTION
+
     def process_manure(self, current_day_conditions: CurrentDayConditions, time: RufasTime) -> dict[str, ManureStream]:
         """Processes manure in Anaerobic Lagoon.
 
@@ -73,7 +78,8 @@ class AnaerobicLagoon(Storage):
         self._manure_to_process = manure_to_return["manure"] if manure_to_return else copy(self.stored_manure)
 
         manure_temperature = self._determine_outdoor_storage_temperature(
-            air_temperature=current_day_conditions.mean_air_temperature
+            air_temperature=current_day_conditions.mean_air_temperature,
+            lower_bound_temperature=ManureConstants.MINIMUM_ANAEROBIC_LAGOON_TEMPERATURE
         )
 
         total_storage_methane, storage_methane_burned = self._apply_methane_emissions(manure_temperature)